Small changes to simplify our codebase.
